[Core/Warrior] Fix MoP rage from white hits and damage taken.
Use stance-based swing rage and Berserker-only damage-taken rage with the MoP 1%/HP formula.

+++ src/server/game/Entities/Unit/Unit.h
@@ -2792,7 +2792,7 @@ public:
     }
     bool IsFalling() const;
 
-    void RewardRage(uint32 baseRage, bool attacker);
+    void RewardRage(float baseRage, bool attacker);
 
     virtual float GetFollowAngle() const
     {
